+++ lib/data/boss_data/nier_boss_class_list.dart
