+++ lib/naer_services/xml_files_randomization/nier_xml_file_randomizer.dart
@@ -1,33 +1,22 @@
 // ignore_for_file: avoid_print
 
 import 'dart:io';
-import 'dart:math';
+import 'package:NAER/naer_services/xml_files_randomization/nier_xml/handle_enemy_groups.dart';
+import 'package:NAER/naer_services/xml_files_randomization/nier_xml/handle_enemy_modification.dart';
 import 'package:NAER/naer_utils/change_tracker.dart';
+import 'package:stack_trace/stack_trace.dart';
 import 'package:xml/xml.dart' as xml;
 import 'package:path/path.dart' as path;
 import 'package:NAER/nier_cli/nier_cli_fork_utils/fileTypeUtils/xml/xmlExtension.dart';
-import 'package:NAER/nier_enemy_data/sorted_data/nier_sorted_enemies.dart';
-import 'package:NAER/nier_enemy_data/boss_data/nier_boss_level_list.dart';
-import 'package:NAER/nier_enemy_data/values_data/nier_important_ids.dart';
+import 'package:NAER/data/sorted_data/nier_sorted_enemies.dart';
+import 'package:NAER/data/boss_data/nier_boss_level_list.dart';
+import 'package:NAER/data/values_data/nier_important_ids.dart';
 import 'package:NAER/naer_services/level_utils/handle_alias_level.dart';
 import 'package:NAER/naer_services/level_utils/handle_boss_level.dart';
 import 'package:NAER/naer_services/level_utils/handle_level.dart';
-import 'package:NAER/naer_services/value_utils/handle_set_type_rtn_flag.dart';
 
 int fileCount = 0;
 int enemyCount = 0;
-
-Random random = Random();
-
-void logMessage(String message) {
-  final logFile = File('log.txt');
-  logFile.writeAsStringSync('$message\n', mode: FileMode.append);
-}
-
-void printAndLog(String message) {
-  print(message);
-  logMessage(message);
-}
 
 Future<void> findEnemiesInDirectory(String directoryPath,
     String sortedEnemiesPath, String enemyLevel, String enemyCategory) async {
@@ -40,39 +29,10 @@ Future<void> findEnemiesInDirectory(String directoryPath,
     sortedEnemyData = enemyData;
   } else {
     // Otherwise, read the sorted enemy data from the file
-    sortedEnemyData = readSortedEnemyData(sortedEnemiesPath);
+    sortedEnemyData = readSortedEnemyDataGroups(sortedEnemiesPath);
   }
 
   await find(directoryPath, sortedEnemyData, enemyLevel, enemyCategory);
-}
-
-Map<String, List<String>> readSortedEnemyData(String filePath) {
-  var file = File(filePath);
-  if (!file.existsSync()) {
-    throw Exception('Sorted enemy data file does not exist: $filePath');
-  }
-
-  try {
-    var content = file.readAsStringSync();
-    var sortedEnemyData = <String, List<String>>{};
-    var matches =
-        RegExp(r'"(\w+)": \[(.*?)\]', multiLine: true).allMatches(content);
-    for (var match in matches) {
-      var group = match.group(1)!;
-      var enemiesStr = match.group(2)!.replaceAll(RegExp(r'[\[\]"]'), '');
-      var enemies = enemiesStr
-          .split(',')
-          .map((e) => e.trim())
-          .where((e) => e.isNotEmpty)
-          .toList();
-      sortedEnemyData[group] = enemies;
-    }
-    print("Sorted Enemy Data: $sortedEnemyData");
-
-    return sortedEnemyData;
-  } catch (e) {
-    throw Exception('Error reading sorted enemy data: $e');
-  }
 }
 
 Future<void> find(
@@ -128,7 +88,7 @@ int traverseDirectory(
   int localFileCount = 0;
   try {
     for (var entity in directory.listSync()) {
-      print('Processing entity: ${entity.path}');
+      // print('Processing entity: ${entity.path}');
       if (entity is File && entity.path.endsWith('.xml')) {
         processXmlFile(entity, sortedEnemyData, enemyLevel, enemyCategory, ids);
         localFileCount++;
@@ -272,36 +232,29 @@ Future<void> processElement(
   }
 }
 
-xml.XmlElement? findValueElementWithName(xml.XmlElement element, String name) {
-  var nameElements = element.findElements('name');
-  if (nameElements.isNotEmpty) {
-    var nameText = nameElements.first.text;
-    if (nameText == name) {
-      return element;
-    }
-  }
-
-  for (var nestedValue in element.findElements('value')) {
-    var result = findValueElementWithName(nestedValue, name);
-    if (result != null) {
-      return result;
-    }
-  }
-
-  return null;
-}
-
+/// This function handles various argument categories and processes the object ID
+/// element accordingly. It checks if the object ID is a boss, handles levels,
+/// and manages different enemy categories.
+///
+/// Parameters:
+/// - [objIdElement]: The XML element containing the object ID.
+/// - [userSelectedEnemyData]: A map of user-selected enemy data grouped by categories.
+/// - [filePath]: The path to the file containing enemy data.
+/// - [enemyLevel]: The level of the enemy.
+/// - [enemyCategory]: The category of the enemy.
+/// - [isImportantId]: A boolean flag indicating if the object ID is important. Defaults to false.
 Future<void> processObjId(
-    xml.XmlElement objIdElement,
-    Map<String, List<String>> userSelectedEnemyData,
-    String filePath,
-    String enemyLevel,
-    String enemyCategory,
-    {bool isImportantId = false}) async {
+  xml.XmlElement objIdElement,
+  Map<String, List<String>> userSelectedEnemyData,
+  String filePath,
+  String enemyLevel,
+  String enemyCategory, {
+  bool isImportantId = false,
+}) async {
   final objIdValue = objIdElement.text;
 
   if (objIdValue.isEmpty) {
-    printAndLog(
+    print(
       'Error: objId value is null or empty in element: ${objIdElement.toXmlString()}',
     );
     return;
@@ -319,220 +272,48 @@ Future<void> processObjId(
 
     switch (enemyCategory) {
       case 'onlybosses':
-        if (isImportantId) {
-          return;
-        } else if (isBossObj) {
+        if (isImportantId) return;
+        if (isBossObj) {
           await handleBossLevel(objIdElement, enemyLevel);
         } else {
           await handleOtherEnemies(
               objIdElement, userSelectedEnemyData, enemyLevel);
         }
         break;
+
       case 'onlyselectedenemies':
-        await handleSelectedEnemies(
+        await handleSelectedObjectIdEnemies(
             objIdElement, userSelectedEnemyData, enemyLevel);
         break;
+
       case 'allenemies':
         if (isBossObj) {
           await handleBossLevel(objIdElement, enemyLevel);
         } else {
-          await handleSelectedEnemies(
+          await handleSelectedObjectIdEnemies(
               objIdElement, userSelectedEnemyData, enemyLevel);
         }
         break;
+
       case 'onlylevel':
         if (isBossObj) {
           await handleBossLevel(objIdElement, enemyLevel);
         } else {
-          await handleOnlyLevel(
+          await handleOnlyObjectIdLevel(
               objIdElement, userSelectedEnemyData, enemyLevel);
         }
         break;
+
       default:
-        if (isImportantId) {
-          return;
-        } else {
-          if (enemyCategory != 'onlylevel') {
-            await handleDefault(objIdElement, userSelectedEnemyData);
-          }
+        if (isImportantId) return;
+        if (enemyCategory != 'onlylevel') {
+          await handleDefaultObjectId(objIdElement, userSelectedEnemyData);
         }
         break;
     }
-  } catch (e) {
-    printAndLog('Error processing $objIdValue: $e');
+  } catch (e, stackTrace) {
+    var trace = Trace.from(stackTrace);
+    print('Error processing $objIdValue: $e');
+    print('Stack trace: ${trace.toString()}');
   }
-}
-
-Future<void> handleOtherEnemies(
-  xml.XmlElement objIdElement,
-  Map<String, List<String>> userSelectedEnemyData,
-  String enemyLevel,
-) async {
-  String? group = findGroupForEmNumber(objIdElement.text, enemyData);
-
-  if (group != null &&
-      !isExcludedGroup(group) &&
-      userSelectedEnemyData[group]?.isNotEmpty == true) {
-    String newEmNumber = userSelectedEnemyData[group]![
-        random.nextInt(userSelectedEnemyData[group]!.length)];
-    replaceTextInXmlElement(objIdElement, newEmNumber);
-    enemyCount++;
-    setSpecificValues(objIdElement, newEmNumber);
-  }
-}
-
-Future<void> handleSelectedEnemies(
-  xml.XmlElement objIdElement,
-  Map<String, List<String>> userSelectedEnemyData,
-  String enemyLevel,
-) async {
-  String? group = findGroupForEmNumber(objIdElement.text, enemyData);
-
-  if (group != null &&
-      !isExcludedGroup(group) &&
-      userSelectedEnemyData[group]?.isNotEmpty == true) {
-    String newEmNumber = userSelectedEnemyData[group]![
-        random.nextInt(userSelectedEnemyData[group]!.length)];
-    replaceTextInXmlElement(objIdElement, newEmNumber);
-    enemyCount++;
-    setSpecificValues(objIdElement, newEmNumber);
-    await handleLevel(objIdElement, enemyLevel, enemyData);
-  }
-}
-
-Future<void> handleOnlyLevel(
-  xml.XmlElement objIdElement,
-  Map<String, List<String>> userSelectedEnemyData,
-  String enemyLevel,
-) async {
-  String? group = findGroupForEmNumber(objIdElement.text, enemyData);
-
-  if (group != null &&
-      !isExcludedGroup(group) &&
-      userSelectedEnemyData[group]?.isNotEmpty == true) {
-    await handleLevel(objIdElement, enemyLevel, enemyData);
-  }
-}
-
-Future<void> handleDefault(
-  xml.XmlElement objIdElement,
-  Map<String, List<String>> userSelectedEnemyData,
-) async {
-  String? group = findGroupForEmNumber(objIdElement.text, enemyData);
-
-  if (group != null &&
-      !isExcludedGroup(group) &&
-      userSelectedEnemyData[group]?.isNotEmpty == true) {
-    String newEmNumber = userSelectedEnemyData[group]![
-        random.nextInt(userSelectedEnemyData[group]!.length)];
-    replaceTextInXmlElement(objIdElement, newEmNumber);
-    enemyCount++;
-    setSpecificValues(objIdElement, newEmNumber);
-  }
-}
-
-bool isExcludedGroup(String group) {
-  const excludedGroups = {'Delete'};
-  return excludedGroups.contains(group);
-}
-
-bool hasAliasOrAncestorAlias(xml.XmlElement objIdElement) {
-  if (objIdElement.getAttribute('alias') != null) {
-    return true;
-  }
-
-  for (var ancestor in objIdElement.ancestors) {
-    if (ancestor is xml.XmlElement && ancestor.getAttribute('alias') != null) {
-      return true;
-    }
-  }
-
-  return false;
-}
-
-bool hasAliasAncestor(xml.XmlElement objIdElement) {
-  return objIdElement.ancestors
-      .any((element) => element.findElements('alias').isNotEmpty);
-}
-
-void updateElement(
-    xml.XmlElement objIdElement, String elementName, String value) {
-  var parentValueElement = objIdElement.parent;
-  if (parentValueElement != null && parentValueElement is xml.XmlElement) {
-    var elements = parentValueElement.findElements(elementName);
-    if (elements.isNotEmpty) {
-      elements.first.children.clear();
-      elements.first.children.add(xml.XmlText(value));
-    }
-  }
-}
-
-dynamic xmlToJson(xml.XmlNode node) {
-  if (node is xml.XmlElement) {
-    final Map<String, dynamic> jsonMap = {};
-
-    if (node.attributes.isNotEmpty) {
-      jsonMap['@attributes'] = {
-        for (var a in node.attributes) a.name.local: a.value
-      };
-    }
-
-    // Handle child nodes
-    for (var child in node.children) {
-      if (child is xml.XmlElement) {
-        var childJson = xmlToJson(child);
-        if (jsonMap.containsKey(child.name.local)) {
-          if (jsonMap[child.name.local] is! List) {
-            jsonMap[child.name.local] = [jsonMap[child.name.local]];
-          }
-          jsonMap[child.name.local].add(childJson);
-        } else {
-          jsonMap[child.name.local] = childJson;
-        }
-      } else if (child is xml.XmlText && child.text.trim().isNotEmpty) {
-        return child.text.trim();
-      }
-    }
-
-    return jsonMap;
-  } else if (node is xml.XmlText && node.text.trim().isNotEmpty) {
-    return node.text.trim();
-  }
-
-  return {};
-}
-
-int compareFilePaths(String a, String b) {
-  var aParts = a.split('/');
-  var bParts = b.split('/');
-
-  int folderCompare = aParts[0].compareTo(bParts[0]);
-  if (folderCompare != 0) {
-    return folderCompare;
-  }
-
-  return aParts[1].compareTo(bParts[1]);
-}
-
-String? findGroupForEmNumber(
-    String emNumber, Map<String, List<String>> enemyData) {
-  for (var group in enemyData.keys) {
-    if (enemyData[group]!.contains(emNumber)) {
-      return group;
-    }
-  }
-  return null;
-}
-
-String getRandomEmNumberFromGroup(
-    String group, Map<String, List<String>> sortedEnemyData) {
-  var groupList = List.from(sortedEnemyData[group]!);
-  groupList.shuffle(); // Shuffle for better randomness
-  var selected = groupList[random.nextInt(groupList.length)];
-  return selected;
-}
-
-void replaceTextInXmlElement(xml.XmlElement element, String newText) {
-  element.children.clear();
-  element.children.add(xml.XmlText(newText));
 }
