+++ lib/naer_ui/animations/shacke_animation_widget.dart
