+++ lib/nier_cli/nier_cli_fork_utils/utils/log_print.dart
@@ -1,0 +1,8 @@
+import 'package:NAER/naer_services/randomize_utils/shared_logs.dart';
+
+final logState = LogState();
+
+void logAndPrint(String message) {
+  print(message);
+  logState.addLog(message);
+}
