+++ pubspec.yaml
@@ -2,11 +2,13 @@ name: NAER
 description: A Tool to Randomize Enemies.
 version: 3.0.0
 
+publish_to: none
+
 environment:
   sdk: '>=3.0.0 <4.0.0'
 
 dev_dependencies:
-  lints: ^3.0.0
+  lints: ^4.0.0
 dependencies:
   archive: ^3.3.6
   args: ^2.4.2
@@ -22,7 +24,7 @@ dependencies:
   open_file: ^3.3.2
   path: ^1.8.3
   path_provider: ^2.1.1
-  flutter_lints: ^3.0.1
+  flutter_lints: ^4.0.0
   uuid: ^4.2.2
   xml: ^6.2.2
   tuple: ^2.0.2
@@ -32,6 +34,7 @@ dependencies:
   crypto: ^3.0.3
   intl: ^0.19.0
   url_launcher: ^6.0.20
+  stack_trace: ^1.11.1
